Remove a couple of memset usages from TSan, introduced in r288624.
TSan runtime shouldn't contain memset, so internal_memset is used
instead and syntax that emits memset is avoided.

This doesn't fail in-tree due to TSan being build with -03, but it fails
when TSan is built with -O0, and is (I think) a true positive.

Patch by Sam McCall, review: https://reviews.llvm.org/D27407

git-svn-id: https://llvm.org/svn/llvm-project/compiler-rt/trunk@288672 91177308-0d34-0410-b5e6-96231b3b80d8